move all svg's into an images folder to make it more compact
